+++ Sources/BraintreeCore/Analytics/BTAnalyticsService.swift
@@ -30,17 +30,23 @@ class BTAnalyticsService: Equatable {
     func sendAnalyticsEvent(
         _ eventName: String,
         correlationID: String? = nil,
+        endpoint: String? = nil,
+        endTime: Int? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,
-        payPalContextID: String? = nil
+        payPalContextID: String? = nil,
+        startTime: Int? = nil
     ) {
         Task(priority: .background) {
             await performEventRequest(
                 eventName,
                 correlationID: correlationID,
+                endpoint: endpoint,
+                endTime: endTime,
                 errorDescription: errorDescription,
                 linkType: linkType,
-                payPalContextID: payPalContextID
+                payPalContextID: payPalContextID,
+                startTime: startTime
             )
         }
     }
@@ -49,17 +55,23 @@ class BTAnalyticsService: Equatable {
     func performEventRequest(
         _ eventName: String,
         correlationID: String? = nil,
+        endpoint: String? = nil,
+        endTime: Int? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,
-        payPalContextID: String? = nil
+        payPalContextID: String? = nil,
+        startTime: Int? = nil
     ) async {
-        let timestampInMilliseconds = Int(round(Date().timeIntervalSince1970 * 1000))
+        let timestampInMilliseconds = Date().utcTimestampMilliseconds
         let event = FPTIBatchData.Event(
             correlationID: correlationID,
+            endpoint: endpoint,
+            endTime: endTime,
             errorDescription: errorDescription,
             eventName: eventName,
             linkType: linkType,
             payPalContextID: payPalContextID,
+            startTime: startTime,
             timestamp: String(timestampInMilliseconds)
         )