+++ Sources/BraintreeCore/BTAPIClient.swift
@@ -2,7 +2,7 @@ import Foundation
 
 /// This class acts as the entry point for accessing the Braintree APIs via common HTTP methods performed on API endpoints.
 /// - Note: It also manages authentication via tokenization key and provides access to a merchant's gateway configuration.
-@objcMembers public class BTAPIClient: NSObject {
+@objcMembers public class BTAPIClient: NSObject, BTHTTPNetworkTiming {
 
     /// :nodoc: This typealias is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     @_documentation(visibility: private)
@@ -18,6 +18,10 @@ import Foundation
 
     /// Client metadata that is used for tracking the client session
     public private(set) var metadata: BTClientMetadata
+
+    /// :nodoc: This property is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
+    @_documentation(visibility: private)
+    public var shouldSendAPIRequestLatency: Bool = false
 
     // MARK: - Internal Properties
     
@@ -65,12 +69,14 @@ import Foundation
 
             tokenizationKey = authorization
             configurationHTTP = BTHTTP(url: baseURL, tokenizationKey: authorization)
+            configurationHTTP?.networkTimingDelegate = self
         case .clientToken:
             do {
                 clientToken = try BTClientToken(clientToken: authorization)
 
                 guard let clientToken else { return nil }
                 configurationHTTP = try BTHTTP(clientToken: clientToken)
+                configurationHTTP?.networkTimingDelegate = self
             } catch {
                 print(errorString + " Missing analytics session metadata - will not send event " + error.localizedDescription)
                 return nil
@@ -470,6 +476,8 @@ import Foundation
             } else if let tokenizationKey, let baseURL {
                 http = BTHTTP(url: baseURL, tokenizationKey: tokenizationKey)
             }
+            
+            http?.networkTimingDelegate = self
         }
 
         if graphQLHTTP == nil {
@@ -480,6 +488,21 @@ import Foundation
             } else if let tokenizationKey, let graphQLBaseURL {
                 graphQLHTTP = BTGraphQLHTTP(url: graphQLBaseURL, tokenizationKey: tokenizationKey)
             }
+        }
+    }
+
+    // MARK: BTAPITimingDelegate conformance
+
+    func fetchAPITiming(path: String, startTime: Int, endTime: Int) {
+        let cleanedPath = path.replacingOccurrences(of: "/merchants/([A-Za-z0-9]+)/client_api", with: "", options: .regularExpression)
+
+        if (shouldSendAPIRequestLatency || path.contains("v1/configuration")) && cleanedPath != "/v1/tracking/batch/events" {
+            analyticsService?.sendAnalyticsEvent(
+                BTCoreAnalytics.apiRequestLatency,
+                endpoint: cleanedPath,
+                endTime: endTime,
+                startTime: startTime
+            )
         }
     }
 }